general settings
added some variables for deciding if a little screen is attached to the pi or not
if buttons are available or not

now you can use this script for 2 things:

	1) /etc/init.d/sysinfoPiTFT

	   run this as endless loop while booting into the runlevel
	   refreshing itself for displaying the latest ip addresses

	   use
		screenAttached=true
		wifiAttached=true 	# if you have usb wifi and wpa_supplicant
	   and
		buttonsAttached=true 	# if you have buttons attached

	2) ~/.bash_profile

	   copy the sysinfoPiTFT script to your home directory with the name .bach_profile
	   every time you log into this pi you'll get this script once
	   without refresh and leaving you into the shell

	   use
		screenAttached=false
		buttonsAttached=false # to avoid the main loop
	   and
		wifiAttached=false # as you desire

added some dependency checks, a little bit of external software is needed... until now its bc, the calculator and figlet to auto-generate the nice big hostname on the tiny little screen. you'll get prompted when - depending on your settings above - installation is needed. currently the "button part" (buttonsAttached=true/false) works with wiringPi and the gpio command (TODO: create native shell solution)